requirements.txt
Flask
gevent
h5py
numpy
Pillow
tensorflow-cpu
Werkzeug
pytest
opencv-python
